pootle_app: Move VCS-specific code in pootle_app.vcs
This simplifies the migration towards Core
