+++ pootle/apps/pootle_translationproject/models.py
@@ -625,8 +625,7 @@ class TranslationProject(models.Model, TreeItem):
 
         from pootle_app.project_tree import (add_files,
                                              match_template_filename,
-                                             direct_language_match_filename,
-                                             sync_from_vcs)
+                                             direct_language_match_filename)
 
         all_files = []
         new_files = []
@@ -644,6 +643,7 @@ class TranslationProject(models.Model, TreeItem):
             file_filter = lambda filename: True
 
         if vcs_sync:
+            from pootle_app.vcs import sync_from_vcs
             sync_from_vcs(ignored_files, ext, self.real_path, file_filter)
 
         all_files, new_files = add_files(
